Adapt iOS classes to updated OfflineRegion type
OfflineRegion in dart now composes around the OfflineRegionDefinition
class rather than inheriting its properties, so this updates the iOS
classes model correspondingly, as well as JSON encoding and decoding.

I had hoped to use Swift's automatic serialization via the `Codable`
protocol, but the loosely typed `metadata` contained in `OfflineRegion`
makes this very difficult (impossible?). This is surprising since an
otherwise strict schema containing a "bag" of untyped data is fairly
common in JSON, but no `Codable` examples I can find deal with it.

That said, the to/from dictionary functions implemented here do the job
in a straightforward way. I rather wish I hadn't wasted so much time
trying to do anything different.

Note that in the dart-side deserializer I made an apparently necessary
change to promote `int` values to `double` for `minZoom` and `maxZoom`.
Without this, I'm seeing an exception when deserializing and I would
expect the same on Android.

+++ lib/src/offline_region.dart
@@ -39,8 +39,9 @@ class OfflineRegionDefinition {
             )
           : null,
       mapStyleUrl: map['mapStyleUrl'],
-      minZoom: map['minZoom'],
-      maxZoom: map['maxZoom'],
+      // small integers may deserialize to Int
+      minZoom: map['minZoom'].toDouble(),
+      maxZoom: map['maxZoom'].toDouble(),
       includeIdeographs: map['includeIdeographs'] ?? false,
     );
   }
@@ -61,8 +62,6 @@ class OfflineRegionDefinition {
     }
     return LatLng(json[0], json[1]);
   }
-
-
 }
 
 /// Description of a downloaded region including its identifier.
